Extract test data into its own module
