+++ io.liquorice.config.formatter.json.gson/build.gradle
@@ -2,5 +2,8 @@ dependencies {
     compile project(':io.liquorice.config.api')
     compile project(':io.liquorice.config.exceptions')
 
+    testCompile project(':io.liquorice.config.test.support')
+
+    // 3rd party dependencies
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.0'
 }
