+++ io.liquorice.config.formatter.json.jackson/build.gradle
@@ -2,6 +2,9 @@ dependencies {
     compile project(':io.liquorice.config.api')
     compile project(':io.liquorice.config.exceptions')
 
+    testCompile project(':io.liquorice.config.test.support')
+
+    // 3rd party dependencies
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.8.5'
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.5'
 }
