+++ io.liquorice.config.formatter.passthrough/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
     compile project(':io.liquorice.config.api')
     compile project(':io.liquorice.config.exceptions')
+
+    testCompile project(':io.liquorice.config.test.support')
 }
