+++ io.liquorice.config.formatter.passthrough/src/test/java/io/liquorice/config/formatter/passthrough/PassThroughConfigFormatterTest.java
@@ -1,11 +1,12 @@
 package io.liquorice.config.formatter.passthrough;
 
-import com.google.common.collect.ImmutableList;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import java.util.List;
 
+import static io.liquorice.config.test.support.ConfigFormatterTestData.TEST_LIST;
+import static io.liquorice.config.test.support.ConfigFormatterTestData.TEST_STRING;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
@@ -14,9 +15,6 @@ import static org.testng.Assert.fail;
  * Created by mthorpe on 12/11/16.
  */
 public class PassThroughConfigFormatterTest {
-
-    private static final String TEST_STRING = "this is a test string";
-    private static final List<String> TEST_LIST = ImmutableList.of("string1", "string2");
 
     private PassThroughConfigFormatter configFormatter;
