+++ io.liquorice.config.storage.file.properties/build.gradle
@@ -3,4 +3,5 @@ dependencies {
     compile project(':io.liquorice.config.exceptions')
 
     testCompile project(':io.liquorice.config.formatter.passthrough')
+    testCompile project(':io.liquorice.config.test.support')
 }
