+++ io.liquorice.config.storage.file.properties/src/test/java/io/liquorice/config/storage/file/properties/WritablePropertiesFileConfigSpaceTest.java
@@ -3,7 +3,6 @@ package io.liquorice.config.storage.file.properties;
 import com.google.common.base.Charsets;
 import io.liquorice.config.api.formatter.ConfigFormatter;
 import io.liquorice.config.formatter.passthrough.PassThroughConfigFormatter;
-import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import java.io.ByteArrayInputStream;
@@ -17,16 +16,21 @@ import java.nio.channels.FileChannel;
 import java.util.Properties;
 import java.util.function.Function;
 
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.BOOL_KEY;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.BOOL_VALUE;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.DOUBLE_KEY;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.DOUBLE_VALUE;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.INT_KEY;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.INT_VALUE;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.LONG_KEY;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.LONG_VALUE;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.STRING_KEY;
-import static io.liquorice.config.storage.file.properties.ConfigSpaceTestData.STRING_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.BOOL_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.BOOL_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DOUBLE_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DOUBLE_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.INT_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.INT_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.LONG_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.LONG_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.STRING_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.STRING_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.UPDATED_BOOL_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.UPDATED_DOUBLE_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.UPDATED_INT_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.UPDATED_LONG_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.UPDATED_STRING_VALUE;
 import static org.mockito.Mockito.mock;
 import static org.testng.Assert.assertEquals;
 
@@ -35,11 +39,34 @@ import static org.testng.Assert.assertEquals;
  */
 public class WritablePropertiesFileConfigSpaceTest {
 
-    private static final boolean UPDATED_BOOL_VALUE = false;
-    private static final double UPDATED_DOUBLE_VALUE = 14.5;
-    private static final int UPDATED_INT_VALUE = 9;
-    private static final long UPDATED_LONG_VALUE = 31L;
-    private static final String UPDATED_STRING_VALUE = "updated string";
+    private static WritablePropertiesFileConfigSpace createConfigSpace(final OutputStream outputStream)
+            throws Exception {
+        // Initialize seed properties
+        final ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        final Properties seedProperties = new Properties();
+        seedProperties.setProperty(BOOL_KEY, Boolean.toString(BOOL_VALUE));
+        seedProperties.setProperty(DOUBLE_KEY, Double.toString(DOUBLE_VALUE));
+        seedProperties.setProperty(INT_KEY, Integer.toString(INT_VALUE));
+        seedProperties.setProperty(LONG_KEY, Long.toString(LONG_VALUE));
+        seedProperties.setProperty(STRING_KEY, STRING_VALUE);
+        seedProperties.store(baos, null);
+        final InputStreamReader isr = new InputStreamReader(new ByteArrayInputStream(baos.toByteArray()),
+                Charsets.UTF_8);
+        final OutputStreamWriter osw = new OutputStreamWriter(outputStream, Charsets.UTF_8);
+
+        final FileChannel mockFileChannel = mock(FileChannel.class);
+        final Function<FileChannel, Reader> fileChannelReaderFunction = internalFileChannel -> isr;
+        final Function<FileChannel, Writer> fileChannelWriterFunction = internalFileChannel -> osw;
+
+        // Initialize sut
+        final ConfigFormatter configFormatter = new PassThroughConfigFormatter();
+        return new WritablePropertiesFileConfigSpace.Builder() //
+                .withConfigFormatter(configFormatter) //
+                .withFileChannel(mockFileChannel) //
+                .withFileChannelReaderFunction(fileChannelReaderFunction) //
+                .withFileChannelWriterFunction(fileChannelWriterFunction) //
+                .build();
+    }
 
     @Test
     public void testUpdateBoolean() throws Exception {
@@ -149,33 +176,5 @@ public class WritablePropertiesFileConfigSpaceTest {
         assertEquals(Integer.parseInt(properties.getProperty(INT_KEY)), INT_VALUE);
         assertEquals(Long.parseLong(properties.getProperty(LONG_KEY)), LONG_VALUE);
         assertEquals(properties.getProperty(STRING_KEY), UPDATED_STRING_VALUE);
-    }
-
-    private static WritablePropertiesFileConfigSpace createConfigSpace(final OutputStream outputStream) throws Exception {
-        // Initialize seed properties
-        final ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        final Properties seedProperties = new Properties();
-        seedProperties.setProperty(BOOL_KEY, Boolean.toString(BOOL_VALUE));
-        seedProperties.setProperty(DOUBLE_KEY, Double.toString(DOUBLE_VALUE));
-        seedProperties.setProperty(INT_KEY, Integer.toString(INT_VALUE));
-        seedProperties.setProperty(LONG_KEY, Long.toString(LONG_VALUE));
-        seedProperties.setProperty(STRING_KEY, STRING_VALUE);
-        seedProperties.store(baos, null);
-        final InputStreamReader isr = new InputStreamReader(new ByteArrayInputStream(baos.toByteArray()),
-                Charsets.UTF_8);
-        final OutputStreamWriter osw = new OutputStreamWriter(outputStream, Charsets.UTF_8);
-
-        final FileChannel mockFileChannel = mock(FileChannel.class);
-        final Function<FileChannel, Reader> fileChannelReaderFunction = internalFileChannel -> isr;
-        final Function<FileChannel, Writer> fileChannelWriterFunction = internalFileChannel -> osw;
-
-        // Initialize sut
-        final ConfigFormatter configFormatter = new PassThroughConfigFormatter();
-        return new WritablePropertiesFileConfigSpace.Builder() //
-                .withConfigFormatter(configFormatter) //
-                .withFileChannel(mockFileChannel) //
-                .withFileChannelReaderFunction(fileChannelReaderFunction) //
-                .withFileChannelWriterFunction(fileChannelWriterFunction) //
-                .build();
     }
 }
