+++ io.liquorice.config.storage.memory/src/test/java/io/liquorice/config/storage/memory/ReadableMapConfigSpaceTest.java
@@ -10,25 +10,25 @@ import org.testng.annotations.Test;
 import java.util.List;
 import java.util.Map;
 
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.BOOL_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.BOOL_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.COMPLEX_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.COMPLEX_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_BOOL_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_COMPLEX_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_DOUBLE_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_INT_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_LONG_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DEFAULT_STRING_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DOES_NOT_EXIST_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DOUBLE_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.DOUBLE_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.INT_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.INT_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.LONG_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.LONG_VALUE;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.STRING_KEY;
-import static io.liquorice.config.storage.memory.ConfigSpaceTestData.STRING_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.BOOL_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.BOOL_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.COMPLEX_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.COMPLEX_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_BOOL_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_COMPLEX_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_DOUBLE_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_INT_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_LONG_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DEFAULT_STRING_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DOES_NOT_EXIST_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DOUBLE_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.DOUBLE_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.INT_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.INT_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.LONG_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.LONG_VALUE;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.STRING_KEY;
+import static io.liquorice.config.test.support.ConfigSpaceTestData.STRING_VALUE;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
