+++ io.liquorice.config.test.support/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    // 3rd party dependencies
+    compile group: 'com.google.guava', name: 'guava', version: '19.0'
+}
