+++ io.liquorice.config.test.support/config/findbugs/findbugs-exclude.xml
@@ -1,0 +1,2 @@
+<FindBugsFilter>
+</FindBugsFilter>
