+++ settings.gradle
@@ -5,3 +5,4 @@ include 'io.liquorice.config.formatter.json.jackson'
 include 'io.liquorice.config.formatter.passthrough'
 include 'io.liquorice.config.storage.file.properties'
 include 'io.liquorice.config.storage.memory'
+include 'io.liquorice.config.test.support'
